Various Changes
I forgot to update my repo for a while
I added electron forge for packaging, and I added filters to make generating random stages more precise. I also tried to use IPC to have my stage info be in it's own file and not chaotically in the render file but completely failed.